Moved all files out of /test/src/default_editor to more appropriate locations
