+++ super_editor/test/super_editor/document_input_ime_test.dart
@@ -5,10 +5,10 @@ import 'package:flutter_test_robots/flutter_test_robots.dart';
 import 'package:super_editor/super_editor.dart';
 import 'package:super_editor/super_editor_test.dart';
 
-import '../../super_editor/document_test_tools.dart';
-import '../../super_editor/test_documents.dart';
-import '../../test_tools.dart';
-import '../_document_test_tools.dart';
+import 'document_test_tools.dart';
+import 'test_documents.dart';
+import '../test_tools.dart';
+import '_document_test_tools.dart';
 
 void main() {
   group('IME input', () {
