+++ super_editor/test/super_editor/infrastructure/common_editor_operations_test.dart
@@ -5,7 +5,7 @@ import 'package:super_editor/super_editor_test.dart';
 import '../_document_test_tools.dart';
 import 'package:flutter_test_robots/flutter_test_robots.dart';
 
-import '../../super_editor/document_test_tools.dart';
+import '../document_test_tools.dart';
 import '../../test_tools.dart';
 
 void main() {
