+++ super_editor/test/super_editor/super_editor_test.dart
@@ -3,7 +3,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:super_editor/super_editor.dart';
 import 'package:super_editor/super_editor_test.dart';
 
-import '../../super_editor/document_test_tools.dart';
+import 'document_test_tools.dart';
 
 void main() {
   group('SuperEditor', () {
