+++ super_editor/test/super_editor/text_entry/super_editor_content_conversion_test.dart
@@ -3,7 +3,7 @@ import 'package:flutter_test_robots/flutter_test_robots.dart';
 import 'package:super_editor/super_editor.dart';
 import 'package:super_editor/super_editor_test.dart';
 
-import '../../super_editor/document_test_tools.dart';
+import '../document_test_tools.dart';
 import '../../test_tools.dart';
 
 void main() {
